+++ Coil/Source/Coil/Utilities/PointerCointainer.h
@@ -9,7 +9,9 @@ namespace Coil
 	class COIL_API PointerContainer
 	{
 	public:
-		PointerContainer() {}
+		PointerContainer()
+		{}
+
 		~PointerContainer()
 		{
 			for (auto element : Container)
@@ -19,10 +21,10 @@ namespace Coil
 		void Push(T* ptr) { Container.push_back(ptr); }
 		T* Pop() { Container.pop_back(); }
 
-		inline typename std::vector<T*>::iterator begin() { return Container.begin();  }
+		inline typename std::vector<T*>::iterator begin() { return Container.begin(); }
 		inline typename std::vector<T*>::iterator end() { return Container.end(); }
 
-		inline int32 size() { return Container.size(); }
+		inline int32 size() { return (int32)Container.size(); }
 
 	private:
 		std::vector<T*> Container;
